requirements.txt
2captcha-python==1.1.3
certifi==2022.9.24
charset-normalizer==2.1.1
et-xmlfile==1.1.0
Faker==15.3.3
greenlet==2.0.1
idna==3.4
openpyxl==3.0.10
playwright==1.28.0
pyee==9.0.4
python-dateutil==2.8.2
python-dotenv==0.21.0
requests==2.28.1
six==1.16.0
transliterate==1.10.2
typing_extensions==4.4.0
urllib3==1.26.13
